Update 0.12
Added input validation for creating an account. Variables affected: (First Name, Last Name, Email, Username, Password)
Usernames are written to a file and this file is read to ensure username is unique
Security flaw - The text file can be edited which can cause issues client-side issues. Most likely a permeant flaw because this is a test application.